+++ core/configureSystem.php
@@ -180,6 +180,8 @@ $max_filesize = "4194304"; //in bytes
 	
 	if(file_exists('/config/.DOCKER') == TRUE){
 		$cfg = '/config/config.php';	
+		
+		symlink($cfg, __ROOT__.'/inc/config.php');
 	}else{
 		$cfg = __ROOT__.'/inc/config.php';
 	}