+++ pages/getFormMeta.php
@@ -33,6 +33,11 @@ while($customers = mysqli_fetch_array($cust)){
     $customer[] = $customers;
 }
 
+$fTypes_q = mysqli_query($conn, "SELECT id,name,description,concentration FROM perfumeTypes ORDER BY id ASC");
+while($fTypes_res = mysqli_fetch_array($fTypes_q)){
+    $fTypes[] = $fTypes_res;
+}
+
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -136,10 +141,9 @@ while($customers = mysqli_fetch_array($cust)){
     <td>
     <select name="finalType" id="finalType" class="form-control ellipsis">  
             <option value="100">Concentrated (100%)</option>
-	 		<option value="<?=$settings['Parfum']?>" <?php if($settings['Parfum'] == $info['finalType']){ echo 'selected';}?>>Parfum (<?=$settings['Parfum']?>%)</option>
-			<option value="<?=$settings['EDP']?>" <?php if($settings['EDP'] == $info['finalType']){ echo 'selected';}?>>EDP (<?=$settings['EDP']?>%)</option>
-	 		<option value="<?=$settings['EDT']?>" <?php if($settings['EDT'] == $info['finalType']){ echo 'selected';}?>>EDT (<?=$settings['EDT']?>%)</option>
-	 		<option value="<?=$settings['EDC']?>" <?php if($settings['EDC'] == $info['finalType']){ echo 'selected';}?>>EDC (<?=$settings['EDC']?>%)</option>		
+	 		<?php foreach ($fTypes as $fType) {?>
+	<option value="<?php echo $fType['concentration'];?>" <?php echo ($info['finalType']==$fType['concentration'])?"selected=\"selected\"":""; ?>><?php echo $fType['name'].' ('.$fType['concentration'];?>%)</option>
+	<?php }	?>	
     </select>
     </td>
   </tr>