+++ pages/todo.php
@@ -1,62 +1,137 @@
-<?php 
-if (!defined('pvault_panel')){ die('Not Found');}
-
-$todo = mysqli_query($conn, "SELECT fid, name, SUM(toadd) AS toAdd FROM makeFormula GROUP BY name ORDER BY name ASC");
-
-?>
-
-<script>
-function removeTODO(fid) {
-	$.ajax({ 
-    url: 'pages/manageFormula.php', 
-	type: 'get',
-    data: {
-		action: 'todo',
-		fid: fid,
-		remove: true,
-		},
-	dataType: 'html',
-    success: function (data) {
-		location.reload();
-	  	$('#msg').html(data);
-    }
-  });
-};
-</script>
 <div id="content-wrapper" class="d-flex flex-column">
-<?php require_once(__ROOT__.'/pages/top.php'); ?>
-        <div class="container-fluid">
-		<?php echo $msg; ?>
-          <div>
-          <div class="card shadow mb-4">
-            <div class="card-header py-3">
-              <h2 class="m-0 font-weight-bold text-primary"><a href="?do=todo">Pending Formulas</a></h2>
-            </div>
-            <div class="card-body">
-              <div class="table-responsive">
-                <table class="table table-bordered" id="tdData" width="100%" cellspacing="0">
-                  <thead>
-                    <tr class="noBorder">
-                      <th colspan="3">
-                      </th>
-                    </tr>
-                    <tr>
-                      <th>Formula Name</th>
-                      <th>Actions</th>
-                    </tr>
-                  </thead>
-                  <tbody id="todo_data">
-                    <?php while ($r = mysqli_fetch_array($todo)) { ?>
-                    <tr>
-                      <td align="center"><a href="pages/makeFormula.php?fid=<?php echo $r['fid']; ?>" target="_blank" class="<?php if($r['toAdd'] == '0'){ echo $class = 'fas fa-check'; } ?>"><?php echo ' '.$r['name']; ?></a></td>
-					  <td align="center"><a href="javascript:removeTODO('<?php echo $r['fid']; ?>')" onclick="return confirm('Delete <?php echo $r['name']; ?>?');" class="fas fa-trash"></a></td>
-					  </tr>
-				  <?php } ?>
-                  </tbody>
-                </table>
-              </div>
-            </div>
-          </div>
-        </div>
+<?php require_once(__ROOT__.'/pages/top.php'); ?>   
+<div class="container-fluid">
+  <div>
+  <div class="card shadow mb-4">
+    <div class="card-header py-3">
+      <h2 class="m-0 font-weight-bold text-primary"><a href="javascript:reload_data()">Pending Formulas</a></h2>
+    </div>
+    <div class="card-body">
+      <div class="table-responsive">
+      <div id="innermsg"></div>
+        <table class="table table-bordered" id="tdDataPending" width="100%" cellspacing="0">
+          <thead>
+            <tr>
+              <th>Formula Name</th>
+              <th>Actions</th>
+            </tr>
+          </thead>
+        </table>
       </div>
     </div>
+  </div>
+</div>
+</div>
+</div>
+<script>
+$(document).ready(function() {
+
+	var tdDataPending = $('#tdDataPending').DataTable( {
+	columnDefs: [
+		{ className: 'pv_vertical_middle text-center', targets: '_all' },
+		{ orderable: false, targets: [1] },
+	],
+	dom: 'lrftip',
+	processing: true,
+	serverSide: true,
+	searching: true,
+	mark: true,
+	language: {
+		loadingRecords: '&nbsp;',
+		processing: 'Please Wait...',
+		zeroRecords: 'No pending formulas found',
+		search: 'Quick Search:',
+		searchPlaceholder: 'Name..',
+		},
+	ajax: {	
+		url: '/core/pending_formulas_data.php?meta=1',
+		type: 'POST',
+		dataType: 'json',
+		data: function(d) {
+				if (d.order.length>0){
+					d.order_by = d.columns[d.order[0].column].data
+					d.order_as = d.order[0].dir
+				}
+			},
+		},
+	   columns: [
+            { data : 'name', title: 'Formula Name', render: name },
+			{ data : null, title: 'Actions', render: actions },
+			],
+	order: [[ 0, 'asc' ]],
+	lengthMenu: [[20, 50, 100, 200, 400], [20, 50, 100, 200, 400]],
+	pageLength: 20,
+	displayLength: 20,
+	});
+	
+}); //END DOC
+
+function name(data, type, row){
+	var css;
+	if (row.toAdd == '0') {
+		 css = 'fas fa-check';
+	}
+	
+	data = '<a href="/pages/makeFormula.php?fid='+ row.fid +'" target="_blank" class="'+ css +'">'+row.name+'</a>' ;
+	
+	return data;
+}
+
+function actions(data, type, row){
+	return '<i rel="tip" title="Delete '+ row.name +'" class="pv_point_gen fas fa-trash" style="color: #c9302c;" id="pend_remove" data-name="'+ row.name +'" data-id='+ row.fid +'></i>';    
+}
+
+function reload_data() {
+    $('#tdDataPending').DataTable().ajax.reload(null, true);
+}
+
+
+$('#tdDataPending').on('click', '[id*=pend_remove]', function () {
+	var frm = {};
+	frm.ID = $(this).attr('data-id');
+	frm.Name = $(this).attr('data-name');
+    
+	bootbox.dialog({
+       title: "Confirm removal",
+       message : 'Remove <strong>'+ frm.Name +'</strong>?',
+       buttons :{
+           main: {
+               label : "Delete",
+               className : "btn-danger",
+               callback: function (){
+			   $.ajax({
+					url: '/pages/manageFormula.php', 
+					type: 'POST',
+					data: {
+						action: 'todo',
+						fid: frm.ID,
+						name: frm.Name,
+						remove: true,
+						},
+					dataType: 'json',
+					success: function (data) {
+						if(data.success) {
+							var msg = '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>' + data.success + '</div>';
+								reload_data();
+							} else {
+								var msg = '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>' + data.error + '</div>';
+				
+							}
+							$('#innermsg').html(msg);
+					}
+				});
+				
+                 return true;
+               }
+           },
+           cancel: {
+               label : "Cancel",
+               className : "btn-default",
+               callback : function() {
+                   return true;
+               }
+           }   
+       },onEscape: function () {return true;}
+   });
+});
+</script>