+++ releasenotes.md
@@ -1,5 +1,8 @@
 Whats New:
 ----------------------
-- PVOnline connectivity improvements
-- Fix a bug preventing set formulas to internal use customer
-- Add stock quantities in the ingredients page
+- Added inventory worth
+- Added last update date in ingredients suppliers
+- Introducing custom perfume types
+- Cart page backend improvements
+- Pending formulas page backend improvements
+- Show ingredient stock/out of stock in search field when adding in formula